system: moved HaskellExternalAnnotator into system
