+++ src/com/haskforce/system/integrations/highlighting/HaskellExternalAnnotator.scala
@@ -1,18 +1,17 @@
-package com.haskforce.haskell.highlighting.annotation.external
+package com.haskforce.system.integrations.highlighting
 
-import scala.collection.JavaConverters._
+import com.haskforce.Implicits._
+import com.haskforce.haskell.highlighting.annotation.external.CompileProblemsProviderFactory
+import com.haskforce.system.integrations.highlighting.HaskellExternalAnnotator.State
+import com.haskforce.tools.hlint.LintProblemsProviderFactory
 import com.intellij.lang.annotation.{AnnotationHolder, ExternalAnnotator}
 import com.intellij.openapi.application.{ApplicationManager, ModalityState}
 import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.fileEditor.FileDocumentManager
 import com.intellij.psi.PsiFile
 import org.jetbrains.annotations.{NotNull, Nullable}
-import com.haskforce.Implicits._
-import com.haskforce.haskell.codeInsight.HaskellCompletionContributor
-import HaskellExternalAnnotator.State
-import com.haskforce.system.integrations.highlighting.{HaskellAnnotationHolder, HaskellProblem, Problems, ProblemsProviderFactory}
-import com.haskforce.system.utils.{SAMUtils, WrappedFuture}
-import com.haskforce.tools.hlint.LintProblemsProviderFactory
+
+import scala.collection.JavaConverters._
 
 /** Single annotator that calls all external tools used for annotations. */
 class HaskellExternalAnnotator extends ExternalAnnotator[PsiFile, State] {
